test: rework failed-trigger rate coverage from review
Move the failed-trigger rate regressions into governance_inv_tests.cpp,
use SetMockTime(0s), and rename ScheduleAdditionalRelay to
ScheduleTriggerRelay since it only applies to triggers.
