+++ src/governance/governance.h
@@ -403,8 +403,8 @@ private:
         EXCLUSIVE_LOCKS_REQUIRED(cs_store);
 
     /** Queue a deferred re-announcement for a trigger that is too new to propagate
-     *  reliably yet. Only call this for objects we are keeping. */
-    void ScheduleAdditionalRelay(const CGovernanceObject& govobj)
+     *  reliably yet. Only call this for triggers we are keeping. */
+    void ScheduleTriggerRelay(const CGovernanceObject& govobj)
         EXCLUSIVE_LOCKS_REQUIRED(cs_store);
 
     bool MasternodeRateCheck(const CGovernanceObject& govobj, bool fUpdateFailStatus, bool fForce, bool& fRateCheckBypassed)
